requirements.txt
autoreject==0.4.3
datasets==3.6.0
einops==0.8.1
janus==2.0.0
matplotlib==3.8.4
mne==1.9.0
numpy==2.2.5
oct2py==5.8.0
opencv_python==4.9.0.80
openpyxl==3.1.5
pandas==2.2.3
picard==2.13.3
Pillow==11.2.1
ptflops==0.7.4
pynvml==12.0.0
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.3
statsmodels==0.14.4
torch==2.4.0+cu124
torcheval==0.0.7
torchinfo==1.8.0
torchvision==0.19.0+cu124
tqdm==4.65.2
transformers==4.40.1
umap==0.1.1
xgboost==3.0.0
